requirements.txt
Django==4.2
DjangoRESTFramework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv
python-dotenv==1.0.0

